+ Elminate warnings exposed by -Wall -Wextra:   - Use unsigned int as for loop counters.   - Comment out unused variables in argument lists.   - Ensure all class data is initilized in the ctor initializer list.   - Ensure data in initializer lists uses the same order that data     list listed in the class declaration.   - Reformat code to match Draco standard indentation and untabify.

+++ src/cdi/test/tDummyEoS.cc
@@ -93,7 +93,7 @@ void test_EoS()
 
     // Retrieve electron based heat capacities.
     std::vector< double > vRefCve( vtemperature.size() );
-    for ( int i=0; i<vtemperature.size(); ++i )
+    for ( size_t i=0; i<vtemperature.size(); ++i )
 	vRefCve[i] = vtemperature[i] + vdensity[i]/1000.0;
 	    
     std::vector< double > vCve = spEoS->getElectronHeatCapacity(
